+++ gradle-plugin/src/main/java/media/barney/crap/gradle/CrapJavaExtension.java
@@ -1,0 +1,8 @@
+package media.barney.crap.gradle;
+
+import org.gradle.api.provider.Property;
+
+public abstract class CrapJavaExtension {
+
+    public abstract Property<Double> getThreshold();
+}